Se cambia el tipo de datos de int a Long para poder obtener el codigo de barras como numero completo. Daba overflow en el caso de INT. Revisar el vto CAE

+++ Modelos/ModeloReporteEncabezadoComprobante.cs
@@ -26,7 +26,7 @@ namespace Modelos
         public string CAINumero { get; set; }
         public DateTime FechaVencimiento { get; set; }
         public string CondicionVenta { get; set; }
-        public int CodigodeBarras { get; set; }
+        public long CodigodeBarras { get; set; }
 
         public ModeloReporteEncabezadoComprobante(ModeloPedido p_mod_pedido) 
         {
